refactor: prepare text extension
