--- src/text/text.h
@@ -1,12 +1,0 @@
-// Copyright (c) 2023 Anton Zhiyanov, MIT License
-// https://github.com/nalgeon/sqlean
-
-// Text handling.
-
-#ifndef TEXT_H
-#define TEXT_H
-
-#include "bstring.h"
-#include "rstring.h"
-
-#endif /* TEXT_H */
